requirements.txt
asgiref==3.2.7
Django==3.0.6
django-crispy-forms==1.9.0
mysqlclient==1.4.6
Pillow==7.1.2
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
